get undo data command working

+++ src/plugins/ac_guiplugin/util/ac_undo.cpp
@@ -17,17 +17,23 @@
 
 #include "ac_undo.h"
 
+#include <ac_imodelitem.h>
 #include <ac_model.h>
 
 #include <mi_idatabase.h>
+
+static bool undoCommandActive = false;
+static bool databaseReading = false;
 
 class UndoCommandPrivate
 {
 public:
     IModelItem *item;
+    quint32 enabled : 32;
 
     UndoCommandPrivate(IModelItem *item)
         :   item(item)
+        ,   enabled(false)
     {}
 
     virtual ~UndoCommandPrivate()
@@ -44,6 +50,16 @@ UndoCommand::~UndoCommand()
     delete d_ptr;
 }
 
+IModelItem *UndoCommand::item() const
+{
+    return d_ptr->item;
+}
+
+void UndoCommand::enable()
+{
+    d_ptr->enabled = true;
+}
+
 class UndoDataCommandPrivate : public UndoCommandPrivate
 {
 public:
@@ -53,22 +69,61 @@ public:
     UndoDataCommandPrivate(IModelItem *item)
         :   UndoCommandPrivate(item)
     {}
+
+    QMap<int, QVariant> data() const
+    {
+        QMap<int, QVariant> map;
+        const int n = item->persistentRoleCount();
+        for (int i = 0;  i < n;  ++i) {
+            const int role = item->persistentRoleAt(i);
+            const QVariant data = item->data(role);
+            map.insert(role, data);
+        }
+        return map;
+    }
+
+    void setData(const QMap<int, QVariant> &data)
+    {
+        const QList<int> keys = data.keys();
+        const int n = keys.count();
+        for (int i = 0;  i < n;  ++i) {
+            const int &role = keys.at(i);
+            item->setData(data.value(role), role);
+        }
+    }
 };
 
 UndoDataCommand::UndoDataCommand(IModelItem *item, QUndoCommand *parent)
     :   UndoCommand(*(new UndoDataCommandPrivate(item)), parent)
-{}
+{
+    Q_D(UndoDataCommand);
+    d->oldData = d->data();
+}
 
 void UndoDataCommand::updateNewData()
 {
+    Q_D(UndoDataCommand);
+    d->newData = d->data();
 }
 
 void UndoDataCommand::redo()
 {
+    if (!d_ptr->enabled)
+        return;
+    Q_D(UndoDataCommand);
+    undoCommandActive = true;
+    d->setData(d->newData);
+    undoCommandActive = false;
 }
 
 void UndoDataCommand::undo()
 {
+    if (!d_ptr->enabled)
+        return;
+    Q_D(UndoDataCommand);
+    undoCommandActive = true;
+    d->setData(d->oldData);
+    undoCommandActive = false;
 }
 
 class UndoListCommandPrivate : public UndoCommandPrivate
@@ -96,13 +151,24 @@ UndoListCommand::~UndoListCommand()
 
 void UndoListCommand::insert()
 {
+    if (!d_ptr->enabled)
+        return;
+    Q_D(UndoListCommand);
+    undoCommandActive = true;
+    d->model()->insertItem(item(), d->row, d->parentIndex);
+    d->deleteItem = false;
+    undoCommandActive = false;
 }
 
 void UndoListCommand::remove()
 {
+    if (!d_ptr->enabled)
+        return;
     Q_D(UndoListCommand);
+    undoCommandActive = true;
     d->model()->takeItem(d->row, d->parentIndex);
     d->deleteItem = true;
+    undoCommandActive = false;
 }
 
 bool UndoListCommand::mergeWith(const QUndoCommand *other)
@@ -124,6 +190,9 @@ class UndoStackPrivate
 public:
     UndoStack *q;
     Model *model;
+    QList<UndoDataCommand*> dataChanges;
+    QList<UndoInsertCommand*> inserts;
+    QList<UndoRemoveCommand*> removes;
 
     UndoStackPrivate(UndoStack *q)
         :   q(q)
@@ -135,7 +204,11 @@ UndoStack::UndoStack(QObject *parent)
     :   QUndoStack(parent)
     ,   d(new UndoStackPrivate(this))
 {
-    Model *model = qobject_cast<Model*>(IDatabase::instance()->model());
+    IDatabase *db = IDatabase::instance();
+    connect(db, SIGNAL(databaseAboutToBeRead()), SLOT(databaseAboutToBeRead()));
+    connect(db, SIGNAL(databaseRead()), SLOT(databaseRead()));
+
+    Model *model = qobject_cast<Model*>(db->model());
     connect(model, SIGNAL(dataAboutToBeChanged(QModelIndex,QModelIndex)), SLOT(dataAboutToBeChanged(QModelIndex,QModelIndex)));
     connect(model, SIGNAL(dataChanged(QModelIndex,QModelIndex)), SLOT(dataChanged(QModelIndex,QModelIndex)));
     connect(model, SIGNAL(modelReset()), SLOT(modelReset()));
@@ -150,46 +223,96 @@ UndoStack::~UndoStack()
     delete d;
 }
 
+void UndoStack::databaseAboutToBeRead()
+{
+    databaseReading = true;
+}
+
+void UndoStack::databaseRead()
+{
+    databaseReading = false;
+}
+
 void UndoStack::dataAboutToBeChanged(const QModelIndex &topLeft, const QModelIndex &bottomRight)
 {
-    Q_UNUSED(topLeft);
     Q_UNUSED(bottomRight);
+    if (undoCommandActive || databaseReading)
+        return;
+    d->dataChanges.append(new UndoDataCommand(d->model->itemFromIndex(topLeft)));
 }
 
 void UndoStack::dataChanged(const QModelIndex &topLeft, const QModelIndex &bottomRight)
 {
-    Q_UNUSED(topLeft);
     Q_UNUSED(bottomRight);
+    if (undoCommandActive || databaseReading)
+        return;
+    IModelItem *item = d->model->itemFromIndex(topLeft);
+    const int n = d->dataChanges.count();
+    for (int i = 0;  i < n;  ++i) {
+        UndoDataCommand *cmd = d->dataChanges.at(i);
+        if (item != cmd->item())
+            continue;
+        cmd->updateNewData();
+        d->dataChanges.removeAt(i);
+        push(cmd);
+        cmd->enable();
+        break;
+    }
 }
 
 void UndoStack::modelReset()
 {
+    clear();
 }
 
 void UndoStack::rowsAboutToBeInserted(const QModelIndex &parent, int start, int end)
 {
-    Q_UNUSED(parent);
-    Q_UNUSED(start);
     Q_UNUSED(end);
+    if (undoCommandActive || databaseReading)
+        return;
+    d->inserts.append(new UndoInsertCommand(d->model->itemFromIndex(d->model->index(start, parent)), start, parent));
 }
 
 void UndoStack::rowsInserted(const QModelIndex &parent, int start, int end)
 {
-    Q_UNUSED(parent);
-    Q_UNUSED(start);
     Q_UNUSED(end);
+    if (undoCommandActive || databaseReading)
+        return;
+    IModelItem *item = d->model->itemFromIndex(d->model->index(start, parent));
+    const int n = d->inserts.count();
+    for (int i = 0;  i < n;  ++i) {
+        UndoInsertCommand *cmd = d->inserts.at(i);
+        if (item != cmd->item())
+            continue;
+        d->inserts.removeAt(i);
+        push(cmd);
+        cmd->enable();
+        break;
+    }
 }
 
 void UndoStack::rowsAboutToBeRemoved(const QModelIndex &parent, int start, int end)
 {
-    Q_UNUSED(parent);
-    Q_UNUSED(start);
     Q_UNUSED(end);
+    if (undoCommandActive || databaseReading)
+        return;
+    d->removes.append(new UndoRemoveCommand(d->model->itemFromIndex(d->model->index(start, parent)), start, parent));
 }
 
 void UndoStack::rowsRemoved(const QModelIndex &parent, int start, int end)
 {
-    Q_UNUSED(parent);
-    Q_UNUSED(start);
     Q_UNUSED(end);
+    if (undoCommandActive || databaseReading)
+        return;
+    IModelItem *item = d->model->itemFromIndex(d->model->index(start, parent));
+    const int n = d->removes.count();
+    for (int i = 0;  i < n;  ++i) {
+        UndoRemoveCommand *cmd = d->removes.at(i);
+        if (item != cmd->item())
+            continue;
+        d->removes.removeAt(i);
+        push(cmd);
+        cmd->enable();
+        break;
+    }
 }
